avoid setting inactive repeatedly because of visibilityChanges

+++ src/hooks/useConnectivityStatus.jsx
@@ -1,8 +1,10 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 
 import { FieldValue } from '../helpers/firebase';
 
 export default function useConnectivityStatus(ref, key = 'inactive') {
+  const [isHidden, setHidden] = useState(document.hidden);
+
   useEffect(() => {
     const setInactive = to => ref.set({
       [key]: to,
@@ -10,7 +12,10 @@ export default function useConnectivityStatus(ref, key = 'inactive') {
 
     // monitor tab changes
     const handleVisibilityChange = () => {
-      setInactive(document.hidden ? FieldValue.serverTimestamp() : FieldValue.delete());
+      if (document.hidden !== isHidden) {
+        setInactive(document.hidden ? FieldValue.serverTimestamp() : FieldValue.delete());
+        setHidden(document.hidden);
+      }
     };
     document.addEventListener('visibilitychange', handleVisibilityChange);
     handleVisibilityChange();
